requirements.txt
PyQt5>=5.15.7
Pillow>=9.0.0
# pyheif>=0.6.0
imageio>=2.22.0
rawpy>=0.18.0
tqdm>=4.64.0
